requirements.txt
antlr4-python3-runtime==4.13.2
asgiref==3.10.0
Django==5.2.8
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
